Fix page height for encounter worksheets & column pages
Giving all sections a max-height of 11in was, unintuitively, causing
documents such as npcs to break to a third column, even when all content
could fit on an 8.5x11 page in two columns. But without max-height for
all sections, encounter worksheets with a lot of content were bleeding
over and causing blank pages. To compromise, I removed the max-height
from section elements in general, but added a max-height of 12in for
encounter worksheets in particular. Why 12 in? Good question; just
seemed to work based on trial and error.